設定値の難読化：PropertyValueCoderの実装クラスを指定可能にする
